Fix CI compile errors and test failures across all ECUs
BSW: Add sys/types.h for ssize_t in test_Can_Posix, fix sign-compare
Integration: Replace TEST_ASSERT_NOT_EQUAL_MESSAGE with valid Unity macro
RZC/BCM: Add STD_TYPES_H guards, fix E_OK/E_NOT_OK macros, fix Dem/Com
  type conflicts in all 12 RZC + 5 BCM test files
TCU: Add DtcStoreEntry_t typedef in test_Swc_DtcStore.c
FZC: Create NvM.h/NvM.c stubs, update test to source-inclusion pattern
CVC: Add SelfTest_Hw_* POSIX stubs in cvc_hw_posix.c
SC: Fix heartbeat confirmation timing, LED blink off-by-one,
  relay CAN bus-off trigger + kill latch persistence

+++ firmware/shared/bsw/mcal/posix/Can_Posix.c
@@ -198,7 +198,7 @@ boolean Can_Hw_Receive(Can_IdType* id, uint8* data, uint8* dlc)
 
     *id  = (Can_IdType)frame.can_id;
     *dlc = frame.can_dlc;
-    if (frame.can_dlc > 8u) {
+    if ((uint32)frame.can_dlc > 8u) {
         *dlc = 8u;
     }
     memcpy(data, frame.data, *dlc);
